Added Discussion components
Modified main_controller to account for discussions. Added the model,
view, and controller for the discussions. Model only contains create
function for now, which is used to create a discussion.

Need to modify discussion view to reflect the proper discussion page.
Also need to add form for discussion creation in the index view (or
wherever seems most appropriate).